pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-ext"
version = "0.1.0"
description = "Extensible framework for Claude Code CLI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["claude", "claude-code", "ai", "agent", "extensions", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.4"]
vault = ["cryptography>=42.0"]
cron = ["croniter>=1.0.0"]
all = ["python-telegram-bot>=21.0", "discord.py>=2.4", "cryptography>=42.0", "croniter>=1.0.0"]
dev = ["pytest>=7.4", "ruff>=0.9.0"]

[project.urls]
Repository = "https://github.com/claudebot101001/claude-ext"

[tool.setuptools.packages.find]
include = ["core*", "extensions*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "B008", "SIM105", "SIM108", "SIM115", "RUF006", "RUF012"]

[tool.ruff.lint.isort]
known-first-party = ["core", "extensions"]

[tool.pytest.ini_options]
testpaths = ["tests"]
